+++ R/downloading_tools.R
@@ -17,12 +17,13 @@
 #' \donttest{
 #' hu <- get_huc(sf::st_sfc(sf::st_point(c(-73, 42)), crs = 4326),
 #'                             type = "huc08")
-#'
+#' if(inherits(hu, "sf")) {
 #' (hu <- substr(hu$huc8, 1, 2))
 #'
 #' download_nhdplushr(tempdir(), c(hu, "0203"), download_files = FALSE)
 #'
 #' download_nhdplushr(tempdir(), c(hu, "0203"), download_files = FALSE, archive = TRUE)
+#' }
 #' }
 download_nhdplushr <- function(nhd_dir, hu_list, download_files = TRUE, archive = FALSE) {